.history/requirements_20251118134852.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.44
psycopg2-binary==2.9.9
python-dotenv==1.0.0
